refactor rap environment
