+++ resources/resourcemanager/adp_resource_manager.py
@@ -10,7 +10,7 @@ import matplotlib.pyplot as plt
 import itertools
 
 from resources.callbacks import ProgressBarManager
-from resources.environments.rap.rap_environment import DeanLinRegionalResourceAllocationEnvironment
+from resources.environments.rap.rap_restricted import DeanLinRegionalResourceAllocationEnvironment
 from resources.environments.rap.adp_environment import ADPResourceAllocationEnvironment
 from resources.region import Region
 from resources.callbacks import SavePerformanceOnCheckpoints, SaveOnBestTrainingRewardCallback
